Moved packages and closed buffered readers
